+++ utils/YZEmbed.js
@@ -1,6 +1,13 @@
 const { RichEmbed } = require('discord.js');
 
 class YZEmbed extends RichEmbed {
+	/**
+	 * A Discord.RichEmbed with predefined properties.
+	 * @param {string} title The embed's title
+	 * @param {string} description The embed's description
+	 * @param {Discord.Message} triggeringMessage The triggering message (default is null)
+	 * @param {boolean} hasAuthor Shows or not the triggering message's author (default is false)
+	 */
 	constructor(title, description, triggeringMessage = null, hasAuthor = false) {
 		super({
 			color: 0x1AA29B,
@@ -11,7 +18,10 @@ class YZEmbed extends RichEmbed {
 		if (triggeringMessage) {
 
 			if (triggeringMessage.channel.type === 'text') {
-				this.setColor(triggeringMessage.member.colorRole.color);
+
+				if (triggeringMessage.member.displayColor) {
+					this.setColor(triggeringMessage.member.displayColor);
+				}
 			}
 
 			if (hasAuthor) {