+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2019-02-24
+### Added
+- New command: `!admin` - Performs bot's maintenance. Only available for the bot's owner.
+
+### Fixed
+- A bug introduced with the last release. Member with no role could'nt execute bot's commands.
+- A thrown error when pushing rolls in DMs.
+
 ## [1.2.0] - 2019-02-23
 ### Added
 - New command: `!rumor` - Tells a random rumor.